implement Food & Beverages grid
