+++ themes/hotel-management-system/hotel-management-theme/src/components/food/index.js
@@ -1,60 +1,40 @@
 import React from "react"
 import Grid from "@material-ui/core/Grid"
 import Typography from "@material-ui/core/Typography"
-import Divider from "@material-ui/core/Divider"
-import Button from "@material-ui/core/Button"
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
-import { faCheck, faXmark } from "@fortawesome/free-solid-svg-icons"
+import FoodCard from "./foodCard"
 import Styles from "./styles"
+import FoodList from "../../../assets/data/foodList"
 
-const FoodCard = ({ imgUrl, title, description, price, availability }) => {
-  const classes = Styles({ availability })
+const FoodPageComponent = () => {
+  const foodItems = FoodList.map(item => (
+    <Grid item key={item.id}>
+      <FoodCard
+        imgUrl={item.imgUrl}
+        title={item.title}
+        description={item.description}
+        price={item.price}
+        availability={item.availability}
+      />
+    </Grid>
+  ))
+  const classes = Styles()
   return (
-    <Grid container direction="column" classes={{ root: classes.container }}>
-      <Grid item>
-        <img src={imgUrl} className={classes.imgStyle} />
+    <Grid container direction="column" alignItems="center">
+      <Grid item classes={{ root: classes.headerStyle }}>
+        <Typography variant="h3" align="center">
+          Food &<span className={classes.decoratedHeader}> Beverages</span>
+        </Typography>
       </Grid>
-      <Grid container direction="column" classes={{ root: classes.wrapper }}>
-        <Grid item>
-          <Typography align="center">{title}</Typography>
-        </Grid>
-        <Grid item>
-          <Divider classes={{ root: classes.divideStyle }} />
-        </Grid>
-        <Grid item classes={{ root: classes.space }}>
-          <Typography variant="h6">{description}</Typography>
-        </Grid>
-        <Grid item>
-          <Grid
-            container
-            justifyContent="space-between"
-            alignItems="center"
-            classes={{ root: classes.space }}
-          >
-            <Grid item>
-              <Typography classes={{ root: classes.priceStyle }}>
-                ${price}
-              </Typography>
-            </Grid>
-            <Grid item>
-              <Button
-                startIcon={
-                  availability ? (
-                    <FontAwesomeIcon icon={faCheck} />
-                  ) : (
-                    <FontAwesomeIcon icon={faXmark} />
-                  )
-                }
-                classes={{ root: classes.btnStyle }}
-              >
-                {availability ? "Available" : "Unavailable"}
-              </Button>
-            </Grid>
-          </Grid>
-        </Grid>
+      <Grid
+        item
+        container
+        direction="row"
+        classes={{ root: classes.itemContainer }}
+      >
+        {foodItems}
       </Grid>
     </Grid>
   )
 }
 
-export default FoodCard
+export default FoodPageComponent
