+++ themes/hotel-management-system/hotel-management-theme/src/pages/index.js
@@ -3,7 +3,6 @@ import Layout from "../ui/layout"
 import HomeContent from "../components/home"
 
 const HomePage = () => {
-  const classes = Styles()
   return (
     <Layout>
       <HomeContent />
